requirements.txt
# Base de datos y ORM (primero porque otros dependen de esto)
SQLAlchemy>=2.0.23
psycopg2-binary>=2.9.9
alembic>=1.12.0
Flask-Migrate>=4.0.5

# Base Flask y extensiones
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-JWT-Extended==4.3.1
Flask-SocketIO==5.1.1

# Seguridad
cryptography>=41.0.0  # Para encriptación Fernet

# Cache y mensajería (actualizado a versiones estables)
redis>=4.5.1
celery>=5.3.6

# Utilidades (necesarias para varios componentes)
python-dotenv>=1.0.0
requests>=2.26.0
gunicorn>=20.1.0

# AI y análisis de datos (actualizados a versiones compatibles)
scikit-learn>=1.0.2
numpy>=1.21.0
pandas>=1.3.0
joblib>=1.1.0

# Desarrollo y testing (al final porque son herramientas)
black>=23.10.0
flake8>=6.1.0
pytest>=7.4.3
pytest-cov>=2.12.1 